Next update

+++ FarmSimulator/src/items/Food.java
@@ -5,13 +5,13 @@ public class Food extends item {
 	/**
 	 * Float representing the proportional increase in healthiness associated with using the item
 	 */
-	private float healthiness;
+	private double healthiness;
 	
 	/**
 	 * Set the items healthiness level
 	 * @param health
 	 */
-	public void setHealthiness(float health) {
+	public void setHealthiness(double health) {
 		healthiness = health;
 	}
 	
@@ -20,7 +20,7 @@ public class Food extends item {
 	 * Return the associated items healthiness contribution
 	 * @return
 	 */
-	public float getHealthiness() {
+	public double getHealthiness() {
 		return healthiness;
 	}
 }
